refactor(setup): implemented review suggestions
