+++ packages/actions/src/index.ts
@@ -4,6 +4,5 @@ export {
     signInToFirebaseWithGithubToken
 } from "./core/auth/index"
 export { getOpenedCeremonies, getCeremonyCircuits } from "./core/contribute/index"
-export { getBucketName, createS3Bucket, objectExist, multiPartUpload } from './helpers/s3'
-export { getCircuitMetadataFromR1csFile, estimatePoT } from './helpers/utils'
-export { setupCeremony } from './core/setup'
+export { getBucketName, createS3Bucket, objectExist, multiPartUpload } from './helpers/storage'
+export { setupCeremony, getCircuitMetadataFromR1csFile, estimatePoT } from './core/setup'
